pyproject.toml
[tool.poetry]
name = "zarda-api"
version = "0.1.0"
description = ""
authors = ["MaHDiaLaGaB <elmahdi.alagab@gmail.com>"]
readme = "README.md"
packages = [{include = "zarda_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
sqlalchemy = "^1.4.45"
psycopg2 = "^2.9.5"
uvicorn = "^0.20.0"
black = "^22.12.0"
pytest = "^7.2.0"
pip = "^22.3.1"
python-multipart = "^0.0.5"
alembic = "^1.9.2"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
